move url_metadata related functions from pdf.py to pdfminer_processing.py
