+++ unstructured/partition/pdf.py
@@ -12,7 +12,7 @@ from typing import IO, TYPE_CHECKING, Any, Optional, cast
 import numpy as np
 import wrapt
 from pdfminer import psparser
-from pdfminer.layout import LTChar, LTContainer, LTImage, LTItem, LTTextBox
+from pdfminer.layout import LTContainer, LTImage, LTItem, LTTextBox
 from pdfminer.utils import open_filename
 from pi_heif import register_heif_opener
 from PIL import Image as PILImage
@@ -69,7 +69,8 @@ from unstructured.partition.pdf_image.pdf_image_utils import (
 from unstructured.partition.pdf_image.pdfminer_processing import (
     clean_pdfminer_inner_elements,
     get_uris,
-    merge_inferred_with_extracted_layout,
+    merge_inferred_with_extracted_layout, check_annotations_within_element, get_word_bounding_box_from_element,
+    map_bbox_and_index,
 )
 from unstructured.partition.pdf_image.pdfminer_utils import (
     open_pdfminer_pages_generator,
@@ -1074,212 +1075,3 @@ def check_coords_within_boundary(
     ) and (coordinates.points[0][1] > boundary_y_min - (vertical_threshold * line_height))
 
     return x_within_boundary and y_within_boundary
-
-
-def calculate_intersection_area(
-    bbox1: tuple[float, float, float, float],
-    bbox2: tuple[float, float, float, float],
-) -> float:
-    """
-    Calculate the area of intersection between two bounding boxes.
-
-    Args:
-        bbox1 (tuple[float, float, float, float]): The coordinates of the first bounding box
-            in the format (x1, y1, x2, y2).
-        bbox2 (tuple[float, float, float, float]): The coordinates of the second bounding box
-            in the format (x1, y1, x2, y2).
-
-    Returns:
-        float: The area of intersection between the two bounding boxes. If there is no
-        intersection, the function returns 0.0.
-    """
-    x1_1, y1_1, x2_1, y2_1 = bbox1
-    x1_2, y1_2, x2_2, y2_2 = bbox2
-
-    x_intersection = max(x1_1, x1_2)
-    y_intersection = max(y1_1, y1_2)
-    x2_intersection = min(x2_1, x2_2)
-    y2_intersection = min(y2_1, y2_2)
-
-    if x_intersection < x2_intersection and y_intersection < y2_intersection:
-        intersection_area = calculate_bbox_area(
-            (x_intersection, y_intersection, x2_intersection, y2_intersection),
-        )
-        return intersection_area
-    else:
-        return 0.0
-
-
-def calculate_bbox_area(bbox: tuple[float, float, float, float]) -> float:
-    """
-    Calculate the area of a bounding box.
-
-    Args:
-        bbox (tuple[float, float, float, float]): The coordinates of the bounding box
-            in the format (x1, y1, x2, y2).
-
-    Returns:
-        float: The area of the bounding box, computed as the product of its width and height.
-    """
-    x1, y1, x2, y2 = bbox
-    area = (x2 - x1) * (y2 - y1)
-    return area
-
-
-def check_annotations_within_element(
-    annotation_list: list[dict[str, Any]],
-    element_bbox: tuple[float, float, float, float],
-    page_number: int,
-    annotation_threshold: float,
-) -> list[dict[str, Any]]:
-    """
-    Filter annotations that are within or highly overlap with a specified element on a page.
-
-    Args:
-        annotation_list (list[dict[str,Any]]): A list of dictionaries, each containing information
-            about an annotation.
-        element_bbox (tuple[float, float, float, float]): The bounding box coordinates of the
-            specified element in the bbox format (x1, y1, x2, y2).
-        page_number (int): The page number to which the annotations and element belong.
-        annotation_threshold (float, optional): The threshold value (between 0.0 and 1.0)
-            that determines the minimum overlap required for an annotation to be considered
-            within the element. Default is 0.9.
-
-    Returns:
-        list[dict[str,Any]]: A list of dictionaries containing information about annotations
-        that are within or highly overlap with the specified element on the given page, based on
-        the specified threshold.
-    """
-    annotations_within_element = []
-    for annotation in annotation_list:
-        if annotation["page_number"] == page_number:
-            annotation_bbox_size = calculate_bbox_area(annotation["bbox"])
-            if annotation_bbox_size and (
-                calculate_intersection_area(element_bbox, annotation["bbox"]) / annotation_bbox_size
-                > annotation_threshold
-            ):
-                annotations_within_element.append(annotation)
-    return annotations_within_element
-
-
-def get_word_bounding_box_from_element(
-    obj: LTTextBox,
-    height: float,
-) -> tuple[list[LTChar], list[dict[str, Any]]]:
-    """
-    Extracts characters and word bounding boxes from a PDF text element.
-
-    Args:
-        obj (LTTextBox): The PDF text element from which to extract characters and words.
-        height (float): The height of the page in the specified coordinate system.
-
-    Returns:
-        tuple[list[LTChar], list[dict[str,Any]]]: A tuple containing two lists:
-            - list[LTChar]: A list of LTChar objects representing individual characters.
-            - list[dict[str,Any]]]: A list of dictionaries, each containing information about
-                a word, including its text, bounding box, and start index in the element's text.
-    """
-    characters = []
-    words = []
-    text_len = 0
-
-    for text_line in obj:
-        word = ""
-        x1, y1, x2, y2 = None, None, None, None
-        start_index = 0
-        for index, character in enumerate(text_line):
-            if isinstance(character, LTChar):
-                characters.append(character)
-                char = character.get_text()
-
-                if word and not char.strip():
-                    words.append(
-                        {"text": word, "bbox": (x1, y1, x2, y2), "start_index": start_index},
-                    )
-                    word = ""
-                    continue
-
-                # TODO(klaijan) - isalnum() only works with A-Z, a-z and 0-9
-                # will need to switch to some pattern matching once we support more languages
-                if not word:
-                    isalnum = char.isalnum()
-                if word and char.isalnum() != isalnum:
-                    isalnum = char.isalnum()
-                    words.append(
-                        {"text": word, "bbox": (x1, y1, x2, y2), "start_index": start_index},
-                    )
-                    word = ""
-
-                if len(word) == 0:
-                    start_index = text_len + index
-                    x1 = character.x0
-                    y2 = height - character.y0
-                    x2 = character.x1
-                    y1 = height - character.y1
-                else:
-                    x2 = character.x1
-                    y2 = height - character.y0
-
-                word += char
-        text_len += len(text_line)
-    return characters, words
-
-
-def map_bbox_and_index(words: list[dict[str, Any]], annot: dict[str, Any]):
-    """
-    Maps a bounding box annotation to the corresponding text and start index within a list of words.
-
-    Args:
-        words (list[dict[str,Any]]): A list of dictionaries, each containing information about
-            a word, including its text, bounding box, and start index.
-        annot (dict[str,Any]): The annotation dictionary to be mapped, which will be updated with
-        "text" and "start_index" fields.
-
-    Returns:
-        dict: The updated annotation dictionary with "text" representing the mapped text and
-            "start_index" representing the start index of the mapped text in the list of words.
-    """
-    if len(words) == 0:
-        annot["text"] = ""
-        annot["start_index"] = -1
-        return annot
-    distance_from_bbox_start = np.sqrt(
-        (annot["bbox"][0] - np.array([word["bbox"][0] for word in words])) ** 2
-        + (annot["bbox"][1] - np.array([word["bbox"][1] for word in words])) ** 2,
-    )
-    distance_from_bbox_end = np.sqrt(
-        (annot["bbox"][2] - np.array([word["bbox"][2] for word in words])) ** 2
-        + (annot["bbox"][3] - np.array([word["bbox"][3] for word in words])) ** 2,
-    )
-    closest_start = try_argmin(distance_from_bbox_start)
-    closest_end = try_argmin(distance_from_bbox_end)
-
-    # NOTE(klaijan) - get the word from closest start only if the end index comes after start index
-    text = ""
-    if closest_end >= closest_start:
-        for _ in range(closest_start, closest_end + 1):
-            text += " "
-            text += words[_]["text"]
-    else:
-        text = words[closest_start]["text"]
-
-    annot["text"] = text.strip()
-    annot["start_index"] = words[closest_start]["start_index"]
-    return annot
-
-
-def try_argmin(array: np.ndarray) -> int:
-    """
-    Attempt to find the index of the minimum value in a NumPy array.
-
-    Args:
-        array (np.ndarray): The NumPy array in which to find the minimum value's index.
-
-    Returns:
-        int: The index of the minimum value in the array. If the array is empty or an
-        IndexError occurs, it returns -1.
-    """
-    try:
-        return int(np.argmin(array))
-    except IndexError:
-        return -1
